sensors: Try to fix flakiness in iframe-related Generic Sensor tests.
The second test, for focus traversal within same-origin frames, was
occasionally failing, especially in the Mac bots.

The flakiness seems to come from the way the "is_sensor_suspended" command
was implemented in iframe_sensor_handler.html. The sensor in the iframe was
created with the default frequency of 5Hz and, consequently, a period of
200ms. "is_sensor_suspended" tried to detect whether the sensor was
suspended by either receiving a new "reading" event (in which case the
sensor was not suspended), or by reaching a timeout function with a period
of 250ms (200ms from the sensor + a small delay).

In some cases, it looks like the 250ms were not enough and we reached the
timeout function prior to the sensor emitting a new "reading" event. Try to
fix this by using a longer timeout of twice the sensor's period to be more
certain that if we reach that code at least one event should have been
emitted before.

While here, make the code easier to follow by documenting what we are doing
and relying on fewer magic numbers.

Bug: 1073865
Change-Id: I630ad6034f0839c17f11f111ead24a88affc9dd2