Attribute animations should not be reflected by getAttribute(...)
According to SMIL [1], when an attribute is animated, the actual DOM
attribute should not change. So only synchronize attributes after a base
value change has taken place.
Rename related flags and methods to not include "animated", since that
seems a bit misleading.
Add SVGElement::CollectStyleForAnimatedPresentationAttributes() to
handle animations of attributes that are presentation attributes but are
not actually specified on the target element.

[1] https://www.w3.org/TR/2001/REC-smil-animation-20010904/#BasicAnim

Bug: 735820
Change-Id: I2a0858dd383a1a96e7b96451a6fca07084983d57